+++ lib/src/app/repo/crud_repository.dart
@@ -26,7 +26,10 @@ abstract class CRUDRepository<Domain extends BasicDomainObject>
   Domain edit(Domain objectToEdit);
 
   ///Destroy the domain.
-  Domain destroy(Domain objectToDestroy);
+  void destroy(Domain objectToDestroy);
+
+  ///Destroy the domain by it's id.
+  void destroyById(int id);
 
   ///Find the correspondent domain by it's Key Id.
   Domain findBy(int keyId);