+++ lib/src/app/repo/external_repo/crud_repository_external_async.dart
@@ -6,19 +6,22 @@ import 'package:clean_core/clean_core.dart';
 abstract class CRUDRepositoryExternalAsync<Entity extends BasicEntityObject>
     extends AbstractExternalRepo {
   //todo: example
-  ///Create the domain.
+  ///Create the entity.
   Future<Entity> create(Entity newObject);
 
-  ///Edit the domain.
+  ///Edit the entity.
   Future<Entity> edit(Entity objectToEdit);
 
-  ///Destroy the domain.
-  Future<Entity> destroy(Entity objectToDestroy);
+  ///Destroy the entity.
+  Future<void> destroy(Entity objectToDestroy);
 
-  ///Find the correspondent domain by it's Key Id.
+  ///Destroy the entity by it's id.
+  Future<void> destroyById(int id);
+
+  ///Find the correspondent entity by it's Key Id.
   Future<Entity> findBy(int keyId);
 
-  ///Find all domains.
+  ///Find all entities.
   Future<List<Entity>> findAll();
 
   ///Count the amount of domains.
@@ -27,7 +30,7 @@ abstract class CRUDRepositoryExternalAsync<Entity extends BasicEntityObject>
     return findAll().then((value) => value.length);
   }
 
-  void init() {}
+  Future<void> init() async {}
 
-  void dispose() {}
+  Future<void> dispose() async {}
 }